+++ frontend/packages/cli/package.json
@@ -37,11 +37,11 @@
   "scripts": {
     "build": "pnpm run '/^build:.*/'",
     "build:cli": "rollup -c  && pnpm run cp:prism",
-    "build:vite": "node scripts/set-env.mjs vite build",
+    "build:vite": "vite build",
     "command:build": "node ./dist-cli/bin/cli.js erd build --input fixtures/input.schema.rb --format schemarb",
     "cp:prism": "cp ../db-structure/node_modules/@ruby/prism/src/prism.wasm ./dist-cli/bin/prism.wasm",
     "dev": "pnpm command:build && cp dist/schema.json public/ && pnpm run '/^dev:.*/'",
-    "dev:app": "node scripts/set-env.mjs vite",
+    "dev:app": "vite",
     "dev:css": "tcm src --watch",
     "fmt": "pnpm run '/^fmt:.*/'",
     "fmt:biome": "biome check --write --unsafe .",
