+++ frontend/packages/cli/vite-plugins/index.ts
@@ -1,0 +1,1 @@
+export * from './set-env.js'
