+++ frontend/packages/cli/vite.config.ts
@@ -2,6 +2,7 @@ import { rmSync } from 'node:fs'
 import react from '@vitejs/plugin-react'
 import tsconfigPaths from 'vite-tsconfig-paths'
 import { defineConfig } from 'vitest/config'
+import { setEnvPlugin } from './vite-plugins/index.js'
 
 const outDir = 'dist-cli/html'
 
@@ -25,7 +26,7 @@ export default defineConfig({
       ],
     },
   },
-  plugins: [react(), tsconfigPaths()],
+  plugins: [react(), tsconfigPaths(), setEnvPlugin()],
   test: {
     globals: true,
     environment: 'node',
